requirements.txt
numpy>=1.24.0
noise>=1.2.2
Pillow>=10.0.0
matplotlib>=3.7.0
scipy>=1.10.0
jupyter>=1.0.0
ipykernel>=6.25.0
